refactor: MVC 패키지 생성, 입출력 관련 함수 View 클래스로 이동
