+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.controller.RacingGameController;
 import racingcar.util.RandomNumberGenerator;
+import racingcar.view.RacingGameView;
 
 public class Application {
     public static void main(String[] args) {
-        new RacingGame(new RandomNumberGenerator()).start();
+        new RacingGameController(new RandomNumberGenerator(), new RacingGameView()).start();
     }
 }
